Space character at the beginning of core_renderer.php breaks non-HTML output
Hi, a customer using Aardvark noticed that their XLS, CSV and other reports are printed as text in the browser, not as downloaded files.

After exploring many options, I noticed that the bug only occurs with Aardvark, so I supposed that something breaks the output.

The space at the beginnig of core_renderer.php was the responsible. I erased it and the reports worked again.